requirements.txt
langchain
python-dotenv
langchain-text-splitters
langchain-openai
langchain_community
ipykernel
sentence_transformers
langchain_huggingface 
faiss-cpu
pypdf
arxiv
pymupdf
langchain_chroma
beautifulsoup4
langchain_groq
langchain_core
fastapi
langserve
openai
duckduckgo-search
wikipedia
streamlit
validators 
youtube_transcript_api
unstructured

